analysis-server/requirements.txt
python-socketio
python-dotenv
Pillow